Start Labwork Homework Problems

+++ ICT1.002 Basic Programming/lecture4/Average/average_implementation.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
+#include "average_header.h"
 
-float average(float num1, float num2)
+float average (float num1, float num2)
 {
-    return ((float)(num1 + num2) / 2);
+    return ((num1 + num2) / 2.0);
 }
